# docs/swagger.yaml
basePath: /
definitions:
  ai.StreamResponse:
    properties:
      choices:
        items:
          properties:
            finish_reason:
              type: string
            index:
              type: integer
            logprobs:
              type: string
            stop_reason:
              type: string
            text:
              type: string
          type: object
        type: array
      created:
        type: integer
      id:
        type: string
      model:
        type: string
      object:
        type: string
      usage:
        type: string
    type: object
  entities.AchievementType:
    properties:
      achievementTypeID:
        type: integer
      events:
        items:
          $ref: '#/definitions/entities.Event'
        type: array
      name:
        type: string
    type: object
  entities.Company:
    properties:
      company:
        type: string
      companyID:
        type: integer
      description:
        type: string
      events:
        items:
          $ref: '#/definitions/entities.Event'
        type: array
      logo:
        type: string
      products:
        items:
          $ref: '#/definitions/entities.Product'
        type: array
      title:
        type: string
    type: object
  entities.Event:
    properties:
      achievementType:
        $ref: '#/definitions/entities.AchievementType'
      achievementTypeID:
        type: integer
      coin:
        type: number
      company:
        $ref: '#/definitions/entities.Company'
      companyID:
        type: integer
      createdAt:
        type: string
      description:
        type: string
      endDs:
        type: string
      erpID:
        type: integer
      eventID:
        type: integer
      eventName:
        type: string
      eventType:
        type: string
      maxUsers:
        type: integer
      startDs:
        type: string
      status:
        type: string
      title:
        type: string
    type: object
  entities.Product:
    properties:
      avalibility:
        type: integer
      company:
        $ref: '#/definitions/entities.Company'
      companyID:
        type: integer
      description:
        type: string
      image:
        type: string
      name:
        type: string
      price:
        type: number
      productID:
        type: integer
    type: object
  usecase.AdminVisitEventInput:
    properties:
      achievement_type_id:
        type: integer
      eventID:
        type: integer
      userID:
        type: integer
    type: object
  usecase.ClosedEventsOutput:
    properties:
      events:
        items:
          $ref: '#/definitions/entities.Event'
        type: array
    type: object
  usecase.EventWithRegistration:
    properties:
      achievementType:
        $ref: '#/definitions/entities.AchievementType'
      achievementTypeID:
        type: integer
      coin:
        type: number
      company:
        $ref: '#/definitions/entities.Company'
      companyID:
        type: integer
      createdAt:
        type: string
      description:
        type: string
      endDs:
        type: string
      erpID:
        type: integer
      eventID:
        type: integer
      eventName:
        type: string
      eventType:
        type: string
      is_registered:
        type: boolean
      maxUsers:
        type: integer
      startDs:
        type: string
      status:
        type: string
      title:
        type: string
    type: object
  usecase.LLMChatInput:
    properties:
      promnt:
        type: string
    type: object
  usecase.LoginOutput:
    properties:
      token:
        type: string
    type: object
  usecase.UpcomingEventInput:
    properties:
      period:
        enum:
        - today
        - tomorrow
        - week
        - month
        type: string
      userID:
        type: integer
    type: object
  usecase.UpcomingEventList:
    properties:
      events:
        items:
          $ref: '#/definitions/usecase.EventWithRegistration'
        type: array
      total:
        type: integer
    type: object
  usecase.UserMeOutput:
    properties:
      birth_date:
        type: string
      coin:
        type: number
      email:
        type: string
      events:
        items:
          $ref: '#/definitions/entities.Event'
        type: array
      l_surname:
        type: string
      name:
        type: string
      phone:
        type: string
      surname:
        type: string
      user_id:
        type: integer
    type: object
  usecase.VisitEventInput:
    properties:
      eventID:
        type: integer
    type: object
info:
  contact: {}
  title: B8boost API
  version: "1.0"
paths:
  /admin/events/visit:
    post:
      parameters:
      - description: input
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/usecase.AdminVisitEventInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: admin visit event
      tags:
      - event
  /events/archived:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usecase.ClosedEventsOutput'
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: get archived events
      tags:
      - event
  /events/upcoming:
    get:
      parameters:
      - description: input
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/usecase.UpcomingEventInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usecase.UpcomingEventList'
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: get upcoming events
      tags:
      - event
  /events/visit:
    post:
      parameters:
      - description: input
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/usecase.VisitEventInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: visit event
      tags:
      - event
  /llm:
    post:
      parameters:
      - description: input
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/usecase.LLMChatInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ai.StreamResponse'
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: chat with llm
      tags:
      - llm
  /login:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usecase.LoginOutput'
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: login with telegram
      tags:
      - auth
  /users/me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usecase.UserMeOutput'
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: get user me
      tags:
      - user
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"